Update test.py
In order to identify genes that are potential candidates for off-target AID mutations, we first need to identify which genes have significant PolII pausing events.  This script is a first pass at an attempt to plot the TR scores (PolII pausing events at TSS / PolII pausing in gene body) for various genes.